# old_api_spec.yaml
openapi: 3.0.0
info:
  title: AI Life Coach API
  version: 1.0.0
servers:
  - url: https://api.yourapp.com
  - url: http://localhost:8000
paths:
  /coach/message:
    post:
      summary: Send a user message to the coach and receive AI response
      tags:
        - coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachMessageRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Coach response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachMessageResponse'
  /api/chat/history:
    get:
      tags:
        - chat
      summary: Retrieve the list of historical messages.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved history messages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatHistoryMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
    post:
      tags:
        - chat
      summary: Add a new message to the history.
      requestBody:
        description: Historical message to store.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatHistoryMessage'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully added the message to history.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /api/info/models:
    get:
      tags:
        - info
      summary: Retrieve model data.
      responses:
        '200':
          description: Successful response with model data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      type: string
                    description: List of available model names.
                  defaults:
                    type: object
                    additionalProperties:
                      type: string
                    description: Default models for each functionality.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
...
components:
  schemas:
    CoachMessageRequest:
      type: object
      properties:
        message:
          type: string
        mode:
          type: string
          enum: ["text", "voice"]
        isolate_from_context:
          type: boolean
        name_of_model:
          type: string
        pick_model_for_chat_context:
          type: string
        pick_model_for_coach_calculation:
          type: string
        pick_model_for_coach_response:
          type: string
        pick_model_for_message_type_checker:
          type: string
        history_range_for_context:
          type: integer
    CoachMessageResponse:
      type: object
      properties:
        response_text:
          type: string
        response_sound:
          type: string
          format: byte
        voice_usage_cost:
          type: number
        voice_usage_duration:
          type: number
        context_calculation_usage:
          type: number
        tts_usage:
          type: number
        stt_usage:
          type: number
    ChatContextNotes:
      type: object
      properties:
        current_chat_context:
          type: string
          description: after each message current chat context is updated.this is calculated internally using multilayer context
        diff_current_chat_context:
          type: string
          description: shows how chat context changed from previouis messages
        where_the_chat_is_going:
          type: string
          description: this is long avarage filter which shows where the conv is going
    UserData:
      type: object
      properties:
        generic_info:
          type: string
        current_mood:
          type: string
        generic_user_goals:
          type: string
        active_user_goals:
          type: string
    CoachContext:
      type: object
      properties:
        goal:
          type: string
        block:
          type: string
        journal_summary:
          type: string
        preferred_tone:
          type: string
    ChatHistoryMessage:
      type: object
      properties:
        user_id:
          type: integer
        user_name:
          type: string
        user_type:
          type: string
        id:
          type: integer
        message:
          type: string
        message_type:
          type: string
        timestamp:
          type: string
          format: date-time
        userData:
          $ref: '#/components/schemas/UserData'
        chatContextNotes:
          $ref: '#/components/schemas/ChatContextNotes'
        coachContext:
          $ref: '#/components/schemas/CoachContext'
      required:
        - user_id
        - id
        - message
        - timestamp
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key needed to authorize requests.












            
  # /journal/entry:
  #   post:
  #     summary: Submit a new journal entry
  #     tags:
  #       - journal
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/JournalEntrySubmit'
  #     security:
  #       - bearerAuth: []
  #     responses:
  #       '201':
  #         description: Created
          
            
...